refactor: 섹션 이름 기능을 Template 아래로 이동
