+++ backend/src/main/java/reviewme/template/controller/SectionController.java
@@ -6,20 +6,20 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RestController;
 import reviewme.reviewgroup.controller.ReviewGroupSession;
 import reviewme.reviewgroup.domain.ReviewGroup;
-import reviewme.template.service.SectionService;
+import reviewme.template.service.TemplateService;
 import reviewme.template.service.dto.response.SectionNamesResponse;
 
 @RestController
 @RequiredArgsConstructor
 public class SectionController {
 
-    private final SectionService sectionService;
+    private final TemplateService templateService;
 
     @GetMapping("/v2/sections")
     public ResponseEntity<SectionNamesResponse> getSectionNames(
             @ReviewGroupSession ReviewGroup reviewGroup
     ) {
-        SectionNamesResponse sectionNames = sectionService.getSectionNames(reviewGroup);
+        SectionNamesResponse sectionNames = templateService.getSectionNames(reviewGroup);
         return ResponseEntity.ok(sectionNames);
     }
 }
