+++ backend/src/test/java/reviewme/api/ApiTest.java
@@ -40,7 +40,6 @@ import reviewme.reviewgroup.service.ReviewGroupLookupService;
 import reviewme.reviewgroup.service.ReviewGroupService;
 import reviewme.template.controller.SectionController;
 import reviewme.template.controller.TemplateController;
-import reviewme.template.service.SectionService;
 import reviewme.template.service.TemplateService;
 
 @WebMvcTest({
@@ -75,9 +74,6 @@ public abstract class ApiTest {
 
     @MockBean
     protected ReviewSummaryService reviewSummaryService;
-
-    @MockBean
-    protected SectionService sectionService;
 
     @MockBean
     protected ReviewGatheredLookupService reviewGatheredLookupService;
