+++ backend/src/test/java/reviewme/fixture/ReviewGroupFixture.java
@@ -4,6 +4,10 @@ import reviewme.reviewgroup.domain.ReviewGroup;
 
 public class ReviewGroupFixture {
 
+    public static ReviewGroup 템플릿_지정_리뷰_그룹(long templateId) {
+        return new ReviewGroup("reviewee", "project", "requestCode", "accessCode", templateId);
+    }
+
     public static ReviewGroup 리뷰_그룹() {
         return 리뷰_그룹("reviewRequestCode", "groupAccessCode");
     }
